+++ .github/workflows/periodic_tests.yml
@@ -109,7 +109,7 @@ jobs:
         run: python -m pip list
 
       - name: Tests
-        run: python -m pytest
+        run: python -m pytest -n logical
 
       - name: Save new cache
         uses: actions/cache/save@v4
@@ -141,7 +141,7 @@ jobs:
         run: python -m pip list
 
       - name: Run tests
-        run: python -m pytest --cov=aeon --cov-report=xml --timeout 1800
+        run: python -m pytest -n logical --cov=aeon --cov-report=xml --timeout 1800
 
       - uses: codecov/codecov-action@v4
         env:
