+++ tsml_eval/evaluation/storage/classifier_results.py
@@ -11,7 +11,7 @@ from sklearn.metrics import (
 )
 
 from tsml_eval.evaluation.storage.estimator_results import EstimatorResults
-from tsml_eval.utils.experiments import write_classification_results
+from tsml_eval.utils.results_writing import write_classification_results
 
 
 class ClassifierResults(EstimatorResults):
